+++ packages/arch3-core/src/archwayclient.ts
@@ -10,7 +10,9 @@ import {
   QueryOutstandingRewardsRequest,
   QueryOutstandingRewardsResponse,
   QueryParamsRequest,
-  QueryParamsResponse
+  QueryParamsResponse,
+  QueryRewardsPoolRequest,
+  QueryRewardsPoolResponse
 } from '@archwayhq/arch3-proto/src/codegen/archway/rewards/v1beta1/query';
 import { CosmWasmClient, HttpEndpoint, SigningCosmWasmClient, SigningCosmWasmClientOptions } from '@cosmjs/cosmwasm-stargate';
 import { OfflineSigner } from "@cosmjs/proto-signing";
@@ -72,6 +74,10 @@ export class ArchwayClient extends CosmWasmClient {
 
   public static async getParams(request: QueryParamsRequest | undefined): Promise<QueryParamsResponse> {
     return await ArchwayClient.rpcQueryClient.archway.rewards.v1beta1.params(request);
+  }
+
+  public static async getPool(request: QueryRewardsPoolRequest | undefined): Promise<QueryRewardsPoolResponse> {
+    return await ArchwayClient.rpcQueryClient.archway.rewards.v1beta1.rewardsPool(request);
   }
 }