+++ changelog.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+
+## Unreleased - 2023-03-17
+ 
+This change adds a getter function for rewards pool.
+ 
+### Added
+- *arch3-proto*: added getter function for rewards pool [#46](https://github.com/archway-network/arch3.js)
+ 
+### Changed
+ 
+### Fixed
+
+
 ## Unreleased - 2023-03-17
  
 This change adds a getter function for params.